[Refactoring]: IndexView 분리 - Onboarding 내부 SubView로 이동
